CuentaBancaria/apps.py
from django.apps import AppConfig


class CuentabancariaConfig(AppConfig):
    default_auto_field = "django.db.models.BigAutoField"
    name = "CuentaBancaria"

    def ready(self):
        import CuentaBancaria.signals
