pyproject.toml
[project]
name = "python-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "social-auth-app-django>=5.7.0",
    "social-auth-app-django-mongoengine>=1.0.0",
]
